add codeSnip.
