--- migration/v4/model/rulebundle.go
@@ -1,17 +1,0 @@
-package model
-
-//
-// RuleBundle - Analysis rules.
-type RuleBundle struct {
-	Model
-	Kind        string
-	Name        string `gorm:"uniqueIndex;not null"`
-	Description string
-	Custom      bool
-	Repository  JSON `gorm:"type:json"`
-	ImageID     uint `gorm:"index" ref:"file"`
-	Image       *File
-	IdentityID  *uint `gorm:"index"`
-	Identity    *Identity
-	RuleSets    []RuleSet `gorm:"constraint:OnDelete:CASCADE"`
-}
